Moved .editconfig and fixed the Code's formate
After Stuart pointed it out the formating when he was reviewing my code; I noticed that put the .editconfig in the wrong folder and not in root. so I made the moved it to root to fix this formating issue. I reformated everthing with the built in tool again. I doubled check this time the file were  formated properly 🤣.