Todo WEB API created In this commit I followed the tutorial which is the first Link which creates a todo list api which uses controllers. I also edited the code from the tutorial to make it use sqlite Database and not a in memory database so that items would still exist after api restarts/reload
Note:
This current version of the api does the CQRS Pattern by it Does have all the crud opertations.  The research already done on how to make the API use the CQRS Design Pattern. So in the next commit i will be changing to use the CQRS Design Pattern.

Links
1. https://learn.microsoft.com/en-us/aspnet/core/tutorials/first-web-api?view=aspnetcore-8.0&tabs=visual-studio-code