APi Converted to Minimal Web API
for the unit testing I had change it from a controller-based api to a Minimal web API because I couldn't find a solution to uses NUnit for unit testing. I die inside thinking about this change. Anyways I had to remove all the logic from the controller and put it into the program.cs and remove lines of code that was used to setup the controllers