NUnit Tests Fix & Improvements
I Added a OneTimeTearDown method and changed SetUp method to OneTimeSetUp method. Then I remove creation of WebApplicationFactory<Program> and HttpClient from being created in each individual test cash and them to be created in the Setup and to be to Dispose on OneTimeTearDown.
Made code smaller by using fluentassertions in the tests.

Bug Fixes:
- Deleting the error fixed from create_and_delete_Todo_Item test.
- Tests failing when ran all at once.