Angular Project Setup/Update/Organized
I create a standalone project and covered the NGModule based project into a standalone one. I did this because standalone application are easier and is the new standard. I decided that I should have a create myself a set of guidlines/file structure plain to keep my code orginized and consistent.I conducted research to gather ideas on how Angular projects are typically organized. I found various designs, but they were mostly similar and so I took all the similar and shared attributes and implemented them in away that made sense for this project. You can read about my decision about the structure in the ReadME.md  I then added bootstrap, Angular Material, Fontawesome, and bootstrap-icons for when it comes time to design the website. Some of these library if not used will be removed later.