First Commit for the interface
Right now I got a partly working interface. I have Designed took a navbar and foot then took it build my own custom drawer and navbar using bootstrap and my own css. I added a Dark mode and light mode them and a switcher. This switcher is on the the navbar for now. I have to do list displaying, and i have 2 buttons that are for deleting and editing the todo lists